# lib/api-zod/src/generated/types/deployment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JuneXDeployment API — Discord bot hosting platform
 * OpenAPI spec version: 0.1.0
 */
import type { DeploymentEnvVars } from './deploymentEnvVars';
import type { DeploymentStatus } from './deploymentStatus';

export interface Deployment {
  id: number;
  userId: number;
  templateId: number;
  templateName: string;
  botName: string;
  /** @nullable */
  herokuAppId?: string | null;
  status: DeploymentStatus;
  createdAt: string;
  updatedAt?: string;
  /** @nullable */
  templateThumbnail?: string | null;
  envVars?: DeploymentEnvVars;
}